タスクが空の時にチュートリアルを表示するようにした
